% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sample.R
\name{analyse_sample}
\alias{analyse_sample}
\title{Classify mutations in a sample.}
\usage{
analyse_sample(
  data,
  sample_name,
  alpha_level = 0.01,
  model = "Binomial",
  rho = NA
)
}
\arguments{
\item{data}{A tibble containing mutations with sample name (sample), gene name
(gene), number of reads with variant (nv), coverage (dp), variant allele
frequency (VAF), and sample purity (purity) as columns.}

\item{sample_name}{The name of the sample to be analysed.}

\item{alpha_level}{The significance level to be used in hypothesis testing.}

\item{model}{The expected distribution for the number of reads with variant
at fixed coverage and purity, that can be chosen between Binomial (no over-dispersion)
or Beta-Binomial (over-dispersion included).}

\item{rho}{If Beta-Binomial model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the test.}
}
\value{
A list containing the input table reduced to the selected sample, with
the additional 'class' column accounting for the classification, the chosen
model, over-dispersion parameter (NA if the model is not Beta-Binomial) and
significance level used in the test.
}
\description{
Classify mutations in a sample.
}
