% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_test.R
\name{binomial_test}
\alias{binomial_test}
\title{Compute likelihoods and uncertainty..}
\usage{
binomial_test(test, DP, purity, cutoff, model, rho = 0.01, karyotypes)
}
\arguments{
\item{test}{Number of reads with the variant.}

\item{DP}{Sequencing coverage of the mutated genome site.}

\item{purity}{Purity of the sample.}

\item{cutoff}{Likelihood cut-off for class assignment.}

\item{model}{Model used for the classification task, either "Binomial" (no over-dispersion)
or "Beta-Binomial" (over-dispersion included), that will be used as the expected
distribution for the number of reads with variant at fixed coverage and purity.}

\item{rho}{If "Beta-Binomial" model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the classification.}

\item{karyotypes}{Karyotypes to be included among the possible classes.}
}
\value{
A tibble including ploidy, multiplicity, proportion of wt alleles,
uncertainty  of the class assignment, model density.
}
\description{
Compute likelihoods and uncertainty..
}
\examples{
binomial_test(test = 220,
DP = 500,
purity = 0.98,
cutoff = 0.9,
model = "Beta-Binomial",
rho = 0.01,
karyotypes = c("1:0","1:1","2:0","2:1","2:2")
)
}
