% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classification}
\alias{classification}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
classification(x)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'}.}
}
\value{
A table with classified data.
}
\description{
Get classification data for specific selected model.
}
