% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eta_prior.R
\name{compute_eta_prior}
\alias{compute_eta_prior}
\title{Evaluate a prior distribution over the rate of reads per chromosom copy from the data.}
\usage{
compute_eta_prior(x, priors_k_m = priors_k_m)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{priors_k_m}{Pre-computed priors for gene muation total copy number and multiplicity.}
}
\description{
Evaluate a prior distribution over the rate of reads per chromosom copy from the data.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
data(priors_pcawg_hmf)
compute_eta_prior(x = MSK_PAAD_output, priors_k_m = priors_pcawg_hmf)
}
