% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{compute_expectations}
\alias{compute_expectations}
\title{Compute the expectation value of k, m and FAM from the model full posterior distribution.}
\usage{
compute_expectations(x)
}
\arguments{
\item{x}{An object of class INCOMMON.}
}
\value{
The same INCOMMON object with annotated expectation values.
}
\description{
Compute the expectation value of k, m and FAM from the model full posterior distribution.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
compute_expectations(x = MSK_PAAD_output)
}
