% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_fit.R
\name{cox_fit}
\alias{cox_fit}
\title{Fit multivariate Cox regression model based on INCOMMON classes.}
\usage{
cox_fit(x, gene, tumor_type, covariates = c("age", "sex", "tmb"))
}
\arguments{
\item{x}{A list of objects of class \code{'INCOMMON'} containing the classification results for
multiple samples, as produced by using function \code{classify}.}

\item{gene}{The selected gene.}

\item{tumor_type}{The selected tumor type.}

\item{covariates}{Covariates used in the multivariate regression.}
}
\value{
An object of class \code{coxph}.
}
\description{
Fit multivariate Cox regression model based on INCOMMON classes.
}
