% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{Data}
\name{data_MSK}
\alias{data_MSK}
\title{Mutation data of the MSK-MetTropism cohort}
\format{
A data frame with 224890 rows and 21 columns:
}
\source{
MSK-MET at cBioPortal https://www.cbioportal.org/study/summary?id=msk_met_2021
}
\usage{
data(data_MSK)
}
\description{
This table contains read counts data and other information
from the MSK-MetTropism cohort.
}
\examples{
data(data_MSK)
data_MSK
}
\keyword{datasets}
