% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example INCOMMON dataset.}
\format{
A named list with the following components:
\describe{
\item{mutations}{A data frame containing mutations.}
\item{sample}{A string indicating the name of the sample.}
\item{tumor_type}{A string indicating the tumor type.}
\item{purity}{A numeric indicating the fraction of tumor in sample.}
}

The columns in the data frame (\code{mutations}) are:
\describe{
\item{chr}{Chromosome harbouring the mutation.}
\item{from}{Genomic start position of the mutation.}
\item{to}{Genomic end position of the mutation.}
\item{ref}{Reference allele.}
\item{alt}{Alternative allele.}
\item{gene}{Gene harbouring the mutation.}
\item{NV}{Number of reads with the variant}
\item{DP}{Total number of reads (sequencing depth)}
\item{VAF}{Variant allele frequency of the mutation.}
}
}
\source{
MSK-MET at cBioPortal https://www.cbioportal.org/study/summary?id=msk_met_2021
}
\usage{
data(example_data)
}
\description{
Genomic data from sample P-0002081-T01-IM3 of the MSK-MET cohort.
}
\details{
This list consists the following components:
\itemize{
\item \code{mutations}: A data frame containing mutations from sample P-0002081-T01-IM3
of the MSK-MET cohort with columns \code{chr}, \code{from}, \code{to}, \code{ref}, \code{alt}, \code{gene}, \code{NV}, \code{DP} and \code{VAF}.
\item \code{sample}: The name of the sample.
\item \code{tumor_type}: The tumor type of the sample.
\item \code{purity}: The purity of the tumor sample (fraction).
}
}
\examples{
data(example_data)
example_data
\dontrun{
# Create an example data
example <- list(
mutations <- data.frame(
 chr = c("chr1", "chr14", "chr17"),  # Chromosome column
 from = c(16265908, 105246551, 7578503),  # Start position column
 to = c(16265908, 105246551, 7578518),  # End position column
 ref = c("A", "C", "CAGGGCAGGTCTTGGC"),  # Reference allele column
 alt = c("T", "T", "-"),  # Alternate allele column
 gene = c("SPEN", "AKT1", "TP53")  # Gene column
),
  sample = "P-0000004-T01-IM3",
  purity = 0.5
)
}

}
\keyword{datasets}
