% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot suitable for \code{'INCOMMON'} classes.}
\usage{
forest_plot(x, tumor_types = FALSE)
}
\arguments{
\item{x}{An object of class \code{'coxph'}.}
}
\value{
A \code{'ggplot2'} object.
}
\description{
Forest plot for multivariate Cox regression based on INCOMMON classes.
}
