% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_classification}
\alias{gene_classification}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
gene_classification(x, gene_id, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{gene_id}{The name of the gene.}

\item{model}{Model used in the test from which to get classification data.}
}
\value{
A tibble with gene-specific classification data.
}
\description{
Get classification data for specific gene under the selected model.
}
