% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genome_interpreter}
\alias{genome_interpreter}
\title{Visualize frequency distribution of INCOMMON classes.}
\usage{
genome_interpreter(x)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by  function \code{classify}.}
}
\value{
An object or a list of class \code{INCOMMON} with additional columns
in \code{classification}.
}
\description{
Visualize frequency distribution of INCOMMON classes.
}
