% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classes}
\alias{get_classes}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_classes(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used in the test from which to get classification data.}
}
\value{
A tibble.
}
\description{
Get significant classification data for the specified model, if already tested.
}
