% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classifier}
\alias{get_classifier}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_classifier(x, model = NULL)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used in the test from which to get classification data.}
}
\value{
A tibble with classification data.
}
\description{
Get classification data for specific selected model.
}
