% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coord}
\alias{get_coord}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
get_coord(x, gene_name)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{gene_name}{Name of the gene affected by the mutation.}
}
\value{
A list of mutation coordinates.
}
\description{
Get coordinates of mutation(s) mapped on a specified gene.
}
