% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_DP}
\alias{get_DP}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_DP(x, id)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{mutation_id}{Coordinates of mutation in the form of a string
containing \code{chr},\code{from},\code{to},\code{alt},\code{ref} coordinates, colon separated.}
}
\value{
DP of the mutation.
}
\description{
Get coverage for a specific mutation in the sample.
}
