% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene}
\alias{get_gene}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_gene(x, mutation_id)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{mutation_id}{Coordinates of mutation in the form of a string
containing \code{chr},\code{from},\code{to},\code{alt},\code{ref} coordinates, colon separated.}
}
\value{
A tibble with the name of the gene affected by the specified mutation.
}
\description{
Get ID of the gene affected by the specified mutation.
}
