% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_id}
\alias{get_id}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_id(x, gene_name)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{gene_name}{Name of the selected gene.}
}
\value{
ID( of mutation(s).
}
\description{
Get ID of the mutation(s) affecting the specified gene
}
