% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_NV}
\alias{get_NV}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_NV(x, mutation_id)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{id}{Coordinates of mutation in the form of a string
containing \code{chr},\code{from},\code{to},\code{alt},\code{ref} coordinates, colon separated.}

\item{from}{Start position of the mutation.}
}
\value{
NV of the mutation.
}
\description{
Get number of reads with variant for a specific mutation in the sample.
}
