% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_purity_bmix}
\alias{get_purity_bmix}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
get_purity_bmix(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used for purity estimation}
}
\value{
Sample purity as obtained by BMix purity estimation procedure.
}
\description{
Extract the sample purity as estimated using the specified model
}
