% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_purity_estimate}
\alias{get_purity_estimate}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
get_purity_estimate(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used for purity estimation}
}
\value{
Purity estimate data.
}
\description{
Extract data and parameters of purity estimation.
}
