% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reliability}
\alias{get_reliability}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
get_reliability(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used for purity estimation}
}
\value{
Reliability of input purity estimate.
}
\description{
Extract reliability of input sample purity as compared to TAPACLOTH estimate using the specified model.
}
