% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_VAF}
\alias{get_VAF}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
get_VAF(x, mutation_id)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{mutation_id}{Coordinates of mutation in the form of a string
containing \code{chr},\code{from},\code{to},\code{alt},\code{ref} coordinates, colon separated.}
}
\value{
VAF of the mutation.
}
\description{
Get variant allele frequency (VAF) for a specific mutation in the sample.
}
