% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{models_avail}
\alias{models_avail}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
models_avail(x)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}
}
\value{
Model names.
}
\description{
List models used for classification tests.
}
