% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSK_classified}
\alias{MSK_classified}
\title{Data from the MSK-MetTropism cohort classified with INCOMMON}
\format{
An INCOMMON object with genomics, clinical data and classification results.
}
\source{
"msk_classified_with_priors.rds" at Zenodo https://zenodo.org/records/10927218
}
\usage{
data(MSK_classified)
}
\description{
This is an INCOMMON object resulting from the classification of all the example data.
}
\examples{
data(MSK_classified)
MSK_classified
}
\keyword{datasets}
