% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSK_clinical_data}
\alias{MSK_clinical_data}
\title{Genomic data of the MSK-MetTropism cohort}
\format{
A tibble with 25659 rows and 15 columns:
}
\source{
MSK-MET at cBioPortal https://www.cbioportal.org/study/summary?id=msk_met_2021
}
\usage{
data(MSK_clinical_data)
}
\description{
This table contains sample names and corresponding clinical data from the MSK-MetTropism cohort.
}
\examples{
data(MSK_clinical_data)
MSK_clinical_data
}
\keyword{datasets}
