% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{Data}
\name{MSK_genomic_data}
\alias{MSK_genomic_data}
\title{Genomic data of the MSK-MetTropism cohort}
\format{
A tibble with 224939 rows and 10 columns:
}
\source{
MSK-MET at cBioPortal https://www.cbioportal.org/study/summary?id=msk_met_2021
}
\usage{
data(MSK_genomic_data)
}
\description{
This table contains sample names and corresponding read counts data from the MSK-MetTropism cohort.
}
\examples{
data(MSK_genomic_data)
MSK_genomic_data
}
\keyword{datasets}
