% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{mutant_dosage_classification}
\alias{mutant_dosage_classification}
\title{Group patients by gene mutant mutant dosage using gene-role specific thresholds.}
\usage{
mutant_dosage_classification(
  x,
  TSG_low = 0.25,
  TSG_high = 0.75,
  ONC_low = 0.33,
  ONC_high = 0.66
)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{TSG_low}{The lower cutoff for mutant dosage classification of tumour suppressor genes.}

\item{TSG_high}{The upper cutoff for mutant dosage classification of tumour suppressor genes.}

\item{ONC_low}{The lower cutoff for mutant dosage classification for oncogenes.}

\item{ONC_high}{The upper cutoff for mutant dosage classification for oncogenes.}
}
\value{
An object of class INCOMMON with new columns reporting mean FAM and class assignment.
}
\description{
Group patients by gene mutant mutant dosage using gene-role specific thresholds.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
mutant_dosage_classification(MSK_PAAD_output, TSG_low = .25, TSG_high = .75, ONC_low = .33, ONC_high = .66)
}
