% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.INCOMMON.R
\name{plot.INCOMMON}
\alias{plot.INCOMMON}
\title{Plot for class \code{'INCOMMON'}.}
\usage{
\method{plot}{INCOMMON}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
If the terzile model is used, the default plot is a histogram of the VAF spectrum is plotted,
with expected clonal peak, purity and VAF of mutations affecting the target gene
highlighted. If the Binomial or Beta-Binomial model is used, it includes
a plot showing details of the used statistical test in the bottom panel.
}
\examples{
input = init(mutations = example_data$data[1,], sample = example_data$sample, purity = example_data$purity)
x = run_classifier(input, alpha_level = 1e-3, model = "Binomial")
x = plot_test(x)
plot.INCOMMON(x)
}
