% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_bmix}
\alias{plot_bmix}
\title{Plot function for class \code{'TAPACLOTH'}.}
\usage{
plot_bmix(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used for purity estimate}
}
\value{
A plot of BMix fit.
}
\description{
Plot results of the BMix fit used for purity estimation, under the specified model.
}
