% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification.R
\name{plot_classification}
\alias{plot_classification}
\title{Visualize classification results for individual observations.}
\usage{
plot_classification(x, sample, assembly = F)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}

\item{sample}{Sample name from the dataset.}

\item{assembly}{Whether to assemble plots of mutations from the same sample in
a multi-faceted plot.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize classification results for individual observations.
}
