% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eta_priors.R
\name{plot_eta_prior}
\alias{plot_eta_prior}
\title{Visualise the prior distribution over the rate of reads per chromosome copy.}
\usage{
plot_eta_prior(priors_eta)
}
\arguments{
\item{priors_eta}{An object of class INCOMMON.}
}
\description{
Visualise the prior distribution over the rate of reads per chromosome copy.
}
\examples{
# Load the default prior on eta, estimated from the MKS-MET data.
data(priors_eta)
# Plot classification results for a specific sample
plot_eta_prior(priors_eta = priors_eta)
}
