% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_gene}
\alias{plot_gene}
\title{#' Getter for class \code{'TAPACLOTH'}.}
\usage{
plot_gene(x, model, gene_name)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{model}{Model used for classification.}

\item{gene_name}{Name of gene affected by the mutation.}
}
\value{
A tibble with classification data for the specified mutation.
}
\description{
Plot results of the classification under the specified model and for the specified gene.
}
