% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plotting function for class \code{'TAPACLOTH'}.}
\usage{
plot_heatmap(x, model = "binomial")
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'} containing a \code{classifier}.}
}
\value{
An object of class \code{'TAPACLOTH'} containing a list of ggplot2
plots named \code{plot_test} inside \code{classifier}.
}
\description{
Generates a composite heatmap with useful information on data and classification results.
}
\examples{
x = init(mutations = example_data$data, sample = example_data$sample, purity = example_data$purity)
x = run_classifier(x)
plot_heatmap(x)
}
