% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{plot_likelihood}
\alias{plot_likelihood}
\title{Plot model likelihood.}
\usage{
plot_likelihood(data, id)
}
\arguments{
\item{data}{An object of class \code{'INCOMMON'}.}

\item{id}{The id of a mutation iin the form sample:chr:from:to:ref:alt:NV:DP.}
}
\value{
A ggplot object.
}
\description{
Plot model likelihood.
}
