% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_k_m.R
\name{plot_posterior_k_m}
\alias{plot_posterior_k_m}
\title{Visualise the posterior distribution on (k,m) configurations.}
\usage{
plot_posterior_k_m(x, k_max = NULL, z_km = NULL)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{k_max}{The maximum allowed value of total copyu number.}

\item{z_km}{The list of posterior distributions over (k,m) configurations.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualise the posterior distribution on (k,m) configurations.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Plot classification results for a specific sample
x = subset_sample(x = MSK_PAAD_output, sample_list = "P-0000142")
plot_posterior_k_m(x = x)
}
