% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_purity_prior.R
\name{plot_purity_prior}
\alias{plot_purity_prior}
\title{Visualise the prior distribution on sample purity}
\usage{
plot_purity_prior(x, sample, purity_error = 0.05)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{sample}{An identifier of the tumour sample.}

\item{purity_error}{The variance of the Beta prior distribution on purity.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualise the prior distribution on sample purity
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Plot classification results for a specific sample
plot_purity_prior(x = MSK_PAAD_output, sample = "P-0000142-T01-IM3", purity_error = 0.05)
}
