% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report.R
\name{plot_report}
\alias{plot_report}
\title{Generate report of the fit}
\usage{
plot_report(x)
}
\arguments{
\item{x}{An object of class INCOMMON.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Generate report of the fit
}
