% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plotting function for class \code{'TAPACLOTH'}.}
\usage{
plot_summary(
  x,
  model = "binomial",
  out_file = "./tapacloth_pheatmap.pdf",
  main = "summary statistics of classification",
  n_frequency_colors = 4
)
}
\arguments{
\item{x}{A list of obj of class \code{'TAPACLOTH'} containing a \code{classifier}.}

\item{model}{Model used for the classification, either "Binomial" or "Beta-Binomial".}
}
\value{
An object of class \code{'pheatmap'}.
}
\description{
Generates a composite heatmap with summary statistics of classification over multiple samples.
}
\examples{
x = init(mutations = example_data$data, sample = example_data$sample, purity = example_data$purity)
x = run_classifier(x)
plot_heatmap(x)
}
