% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test.R
\name{plot_test}
\alias{plot_test}
\title{#'Plotting function for class \code{'TAPACLOTH'}.}
\usage{
plot_test(x)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}
}
\value{
An object of class \code{'TAPACLOTH'} containing a list of ggplot2
plots named \code{plot_test} inside \code{classifier}.
}
\description{
Produces a list of plots, one for each mutation in the input, displaying the
results of classification.
}
