% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tropism.R
\name{plot_tropism}
\alias{plot_tropism}
\title{Visualize metastatic propnesity odds ratio in a volcano plot fashion.}
\usage{
plot_tropism(x, tumor_type)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}

\item{tumor_type}{The tumour type for which classified data is available.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize metastatic propnesity odds ratio in a volcano plot fashion.
}
