% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior}
\alias{posterior}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
posterior(x, id)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}
}
\value{
A tibble containing parameters for all the models used in the classification.
}
\description{
Get model parameters of the performed classification tests.
}
