% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{posterior_k_m}
\alias{posterior_k_m}
\title{Get the posterior distribution over (k,m) configurations for a specific mutation}
\usage{
posterior_k_m(x, id)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{id}{The id of the mutation (sample:chr:from:to:ref:alt:NV:DP)}
}
\value{
A table with the estimated posterior distribution.
}
\description{
Get the posterior distribution over (k,m) configurations for a specific mutation
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
show_FAM(MSK_PAAD_output, tumor_type = 'PAAD', gene = 'TP53')
}
