% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.INCOMMON.R
\name{print.INCOMMON}
\alias{print.INCOMMON}
\title{Print for class \code{'INCOMMON'}.}
\usage{
\method{print}{INCOMMON}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'INCOMMON'}.
}
\examples{
data(MSK_genomic_data)
data(MSK_clinical_data)
sample = 'P-0002081'
x = init(
  genomic_data = MSK_genomic_data[MSK_genomic_data$sample == sample,],
  clinical_data = MSK_clinical_data[MSK_clinical_data$sample == sample,]
)
print(x)
}
