% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tapacloth.R
\name{print.TAPACLOTH}
\alias{print.TAPACLOTH}
\title{Print for class \code{'TAPACLOTH'}.}
\usage{
\method{print}{TAPACLOTH}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'TAPACLOTH'}.
}
