% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priors_pcawg_hmf}
\alias{priors_pcawg_hmf}
\title{Priors over (k,m) configurations from PCAWG and HMF}
\format{
A data frame with 26,280 rows and 6 columns:
\describe{
\item{gene}{Name of the gene (Hugo Symbol).}
\item{tumor_type}{Tumor type.}
\item{k}{Total copy number}
\item{m}{Mutation multiplicity}
\item{N}{Total number of samples}
\item{n}{Number of samples per (k,m) configuration}
}
}
\source{
Validated copy number calls from PCAWG: https://doi.org/10.5281/zenodo.6410935
}
\usage{
data(priors_pcawg_hmf)
}
\description{
Prior distribution of copy number and mutation multiplicity from PCAWG.
}
\examples{
data(priors_pcawg_hmf)
priors_pcawg_hmf
}
\keyword{datasets}
