% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{show_FAM}
\alias{show_FAM}
\title{Get frction of alleles with the mutation (FAM) values for a gene and cancer type.}
\usage{
show_FAM(x, tumor_type = NULL, gene = NULL)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{gene}{The gene name..}

\item{tumour_type}{The tumour type identifier.}
}
\value{
A table with the estimated FAM values
}
\description{
Get frction of alleles with the mutation (FAM) values for a gene and cancer type.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
show_FAM(MSK_PAAD_output, tumor_type = 'BRCA', gene = 'TP53')
}
