% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_new.R
\name{subset_sample}
\alias{subset_sample}
\title{Subset an INCOMMON object by sample ID.}
\usage{
subset_sample(x, sample_list)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} generated with function \code{init}.}

\item{sample_list}{a list of identifiers for the samples to be subsetted}
}
\value{
An object of class \code{INCOMMON} containing a subset of the original input.
}
\description{
Subset an INCOMMON object by sample ID.
}
\examples{
# First load example data
data(MSK_PAAD_output)
subset_sample(x = MSK_PAAD_output, sample_list = c("P-0000142-T01-IM3"))
print(x)
}
