% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classification}
\alias{classification}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
classification(x)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{model}{Model used in the test from which to get classification data.}
}
\value{
A tibble with classification data.
}
\description{
Get classification data for specific selected model.
}
