% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify mutations using a Beta-Binomial model-based test.}
\usage{
classify(
  x,
  priors = pcawg_priors,
  entropy_cutoff = 0.2,
  rho = 0.01,
  parallel = FALSE,
  num_cores = NULL,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2")
)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} generated with function \code{init}.}

\item{priors}{A dplyr::tibble or data frame with columns \code{gene}, \code{tumor_type}, \code{label} and \code{p} indicating tumor-specific
or pan-cancer (PANCA) prior probabilities.}

\item{entropy_cutoff}{Entropy cut-off for Tier-1 vs Tier-2 assignment.}

\item{rho}{Over-dispersion parameter.}

\item{parallel}{Whether to run the classification in parallel (default: FALSE)}

\item{num_cores}{The number of cores to use for parallel classification.
By default, it takes 80\% of the available cores.}

\item{karyotypes}{Karyotypes to be included among the possible classes.}
}
\value{
An object of class \code{INCOMMON} containing the original input plus
the classification data and parameters.
}
\description{
Classify mutations using a Beta-Binomial model-based test.
}
