% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Compute model likelihood, posterior and entropy}
\usage{
compute_posterior(
  NV,
  DP,
  gene,
  priors = NULL,
  tumor_type,
  purity,
  entropy_cutoff,
  rho = 0.01,
  karyotypes,
  assign_extremes
)
}
\arguments{
\item{NV}{Number of reads with the variant.}

\item{DP}{Sequencing coverage of the mutated genome site.}

\item{gene}{Gene name or symbol.}

\item{priors}{Prior distribution.}

\item{tumor_type}{Tumor type of the sample.}

\item{purity}{Purity of the sample.}

\item{entropy_cutoff}{Cut-off on entropy for Tier-1/Tier-2 distinction.}

\item{rho}{If "Beta-Binomial" model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the classification.}

\item{karyotypes}{Karyotypes to be included among the possible classes.}

\item{model}{Model used for the classification task, either "Binomial" (no over-dispersion)
or "Beta-Binomial" (over-dispersion included), that will be used as the expected
distribution for the number of reads with variant at fixed coverage and purity.}
}
\value{
A tibble including ploidy, multiplicity, proportion of wt alleles,
uncertainty  of the class assignment, model density.
}
\description{
Compute model likelihood, posterior and entropy
}
\examples{
binomial_test(test = 220,
DP = 500,
purity = 0.98,
cutoff = 0.9,
model = "Beta-Binomial",
rho = 0.01,
karyotypes = c("1:0","1:1","2:0","2:1","2:2")
)
}
