% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_purity.R
\name{estimate_purity}
\alias{estimate_purity}
\title{Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.}
\usage{
estimate_purity(x, model = "Binomial", eps = 0.01)
}
\arguments{
\item{x}{An object of class \code{'TAPACLOTH'} generated with function \code{init}.}

\item{model}{The expected distribution for the number of reads with variant
at fixed coverage and purity, that can be chosen between Binomial (no over-dispersion)
or Beta-Binomial (over-dispersion included).}

\item{eps}{In case data is fitted with a mixture of 2 distributions, sample purity
is estimated using only that whose peak is closer to the input purity estimate
than the value of this parameter. If both distributions satisfy this condition,
the one which is closest to the input purity estimate is used for the new estimate.}
}
\value{
An object of class \code{TAPACLOTH} containing the inferred purity.
}
\description{
Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.
}
\examples{
input = init(mutations = example_data$data, sample = example_data$sample, purity = example_data$purity)
out = estimate_purity(x = input, model = "binomial", eps = 0.01)
print(out)
}
