% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_purity.R
\name{estimate_purity}
\alias{estimate_purity}
\title{Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.}
\usage{
estimate_purity(data, sample_name, model = "Binomial", purity = 1, eps = 0.01)
}
\arguments{
\item{data}{A tibble containing mutations with sample name (sample), gene name
(gene), number of reads with variant (nv), coverage (dp), variant allele
frequency (VAF), and sample purity (purity) as columns.}

\item{sample_name}{The name of the sample for which to infer purity.}

\item{model}{The expected distribution for the number of reads with variant
at fixed coverage and purity, that can be chosen between Binomial (no over-dispersion)
or Beta-Binomial (over-dispersion included).}

\item{purity}{An input purity estimate.}

\item{eps}{In case data is fitted with a mixture of 2 distributions, sample purity
is estimated using only that whose peak is closer to the input purity estimate
than the value of this parameter. If both distributions satisfy this condition,
the one which is closest to the input purity estimate is used for the new estimate.}
}
\value{
A list containing:
-the input table reduced to the selected sample, with
additional columns \code{purity_bmix} for the inferred purity and \code{purity_error} for
the relative error between the inferred and the input purity, assuming the
inferred value is correct.
-an object of class \code{bmix} that represents a fit mixtur
-the inferred sample purity and
-a cowplot figure showing results.
}
\description{
Infer sample purity using a mixture of Binomial or Beta-Binomial distributions.
}
