% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Getter for class \code{'TAPACLOTH'}.}
\usage{
get_params(x, model)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}
}
\value{
A tibble containing parameters for all the models used in the classification.
}
\description{
Get model parameters of the performed classification tests.
}
