% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{#' Prepare input for analyses with \code{'TAPACLOTH'}.}
\usage{
init(mutations, sample, purity)
}
\arguments{
\item{mutations}{A tibble listing mutations called in the sample
in the form of chromosome \code{chr}, start position \code{from}, end position \code{to},
reference \code{ref} and alternative \code{alt} alleles, gene name \code{gene} as Hugo Symbol.}

\item{sample}{Sample name.}

\item{purity}{Sample purity.}
}
\value{
A list with mutations including gene role as annotated in COSMIC
cancer gene census, sample name and purity, in the format requested for
\code{TAPACLOTH} analyses.
}
\description{
Cast input data in the format requested by \code{'TAPACLOTH'} \code{run_classifier}
and \code{estimate_purity} functions.
}
