% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Prepare input for analyses with \code{'TAPACLOTH'}.}
\usage{
init(mutations, sample, purity, gene_roles = cancer_gene_census)
}
\arguments{
\item{mutations}{a tibble with columns chromosome \code{chr}, start position \code{from}, end position \code{to},
reference \code{ref} and alternative \code{alt} alleles, coverage \code{DP}, number
of reads with variant \code{NV}, variant allelic frequency \code{VAF} gene name \code{gene}
as Hugo Symbol.}

\item{sample}{Sample name.}

\item{purity}{Sample purity.}

\item{gene_roles}{A tibble reporting a list of gene names \code{gene} and associated
role \code{gene_role} among "oncogene", "TSG" and "fusion". Default is taken from COSMIC cancer gene census.}
}
\value{
An object of class \code{TAPACLOTH}.
}
\description{
Process input data into an object of class \code{'TAPACLOTH'}, ready for
analyses with \code{run_classifier} and \code{estimate_purity} functions.
}
