% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier_fit.R
\name{kaplan_meier_fit}
\alias{kaplan_meier_fit}
\title{Fit Kaplan-Meier survival model based on INCOMMON classes.}
\usage{
kaplan_meier_fit(x, tumor_type, gene)
}
\arguments{
\item{x}{A list of objects of class \code{'INCOMMON'} containing the classification results for
multiple samples, as produced by using function \code{classify}.}

\item{tumor_type}{The selected tumor type.}

\item{gene}{The selected gene.}
}
\value{
An table containing the selected gene, tumor_type, the data table used to
fit the model and an object of class \code{'survfit'}.
}
\description{
Fit Kaplan-Meier survival model based on INCOMMON classes.
}
