% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_propensity.R
\name{met_propensity}
\alias{met_propensity}
\title{Logistic regression of metastatic propensity based on INCOMMON classes.}
\usage{
met_propensity(x, gene, tumor_type)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results
as produced by function \code{classify}.}

\item{gene}{The gene on which patient's stratification is based.}

\item{tumor_type}{The tumor type of patients to stratify.}
}
\value{
An object of class \code{'INCOMMON'} containing an additional object \code{survival}.
}
\description{
Logistic regression of metastatic propensity based on INCOMMON classes.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Estimate the metastatic propensity associated with mutant TP53 with vs without CNA in BRCA.
MSK_PAAD_output = met_propensity(x = MSK_PAAD_output, tumor_type = 'BRCA', gene = 'TP53')
}
