% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_tropism.R
\name{met_tropism}
\alias{met_tropism}
\title{Logistic regression of metastatic tropism based on INCOMMON classes.}
\usage{
met_tropism(x, gene, tumor_type, metastatic_site)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results
as produced by function \code{classify}.}

\item{gene}{The gene on which patient's stratification is based.}

\item{tumor_type}{The tumor type of patients to stratify.}

\item{metastic_site}{The target organ of metastatic diffusion.}
}
\value{
An object of class \code{'INCOMMON'} containing an additional object \code{survival}.
Logistic regression of metastatic tropism based on INCOMMON classes.
}
\description{
Logistic regression of metastatic tropism based on INCOMMON classes.
}
