% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameters}
\alias{parameters}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
parameters(x)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}
}
\value{
A dplyr::tibble containing parameters for all the models used in the classification.
}
\description{
Get model parameters of the performed classification tests.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Get classification parameters
parameters(MSK_PAAD_output)
}
