% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tapacloth.R
\name{plot.TAPACLOTH}
\alias{plot.TAPACLOTH}
\title{Plot for class \code{'TAPACLOTH'}.}
\usage{
\method{plot}{TAPACLOTH}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'TAPACLOTH'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
If the terzile model is used, the default plot is a histogram of the VAF spectrum is plotted,
with expected clonal peak, purity and VAF of mutations affecting the target gene
highlighted. If the Binomial or Beta-Binomial model is used, it includes
a plot showing details of the used statistical test in the bottom panel.
}
