% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_fraction.R
\name{plot_class_fraction}
\alias{plot_class_fraction}
\title{Visualize frequency distribution of INCOMMON classes.}
\usage{
plot_class_fraction(x, tumor_type, gene)
}
\arguments{
\item{x}{A list of objects of class \code{'INCOMMON'} containing the classification results, as
produced by using function \code{classify}.}

\item{tumor_type}{Tumor type for tumor-specific prior ('PANCA' for pan-cancer).}

\item{gene}{Gene for gene-specific prior.}
}
\value{
An object or a list of class \code{'ggplot2'}.
}
\description{
Visualize frequency distribution of INCOMMON classes.
}
