% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification.R
\name{plot_classification}
\alias{plot_classification}
\title{Visualize classification results for individual observations.}
\usage{
plot_classification(x, assembly = F)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize classification results for individual observations.
}
\examples{
x = init(mutations = example_data$data,
         sample = example_data$sample,
         purity = example_data$purity,
         tumor_type = example_data$tumor_type)
x = classify(
    x = x, 
    priors = priors,
    entropy_cutoff = 0.2,
    rho = 0.01,
    karyotypes = c("1:0","1:1","2:0","2:1","2:2")
    )
plot_test(x = x, id = ids(x)[1])
}
