% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eta_prior.R
\name{plot_eta_prior}
\alias{plot_eta_prior}
\title{Visualise the prior distribution over the rate of reads per chromosome copy.}
\usage{
plot_eta_prior(x, priors_eta)
}
\arguments{
\item{x}{An object of class INCOMMON.'}

\item{priors_eta}{A table with the prior distribution on the rate of reads per chromosome copy as obtained with function \code{compute_eta_prior}.}
}
\description{
Visualise the prior distribution over the rate of reads per chromosome copy.
}
\examples{
# Load the default prior on eta, estimated from the MKS-MET data and the data.
data(MSK_genomic_data)
data(MSK_clinical_data)
data(priors_eta)
x = init(genomic_data = MSK_genomic_data, clinical_data =MSK_clinical_data)
# Plot classification results for a specific sample
plot_eta_prior(x = x, priors_eta = priors_eta)
}
