% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_volcano.R
\name{plot_met_volcano}
\alias{plot_met_volcano}
\title{Visualize metastatic propnesity odds ratio in a volcano plot fashion.}
\usage{
plot_met_volcano(x, tumor_type)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}

\item{tumor_type}{The tumour type for which classified data is available. If 'PANCA'
it pools from multiple tumour types into a pan-cancer visualisation.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize metastatic propnesity odds ratio in a volcano plot fashion.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Estimate the metastatic propensity associated with mutant KRAS and CDKN2A high vs balanced dosage.
for(g in c('KRAS', 'CDKN2A')){MSK_PAAD_output = met_propensity(x = MSK_PAAD_output, tumor_type = 'PAAD', gene = g)}
# Plot results in a volcano plot
plot_met_volcano(x = MSK_PAAD_output, tumor_type = 'PAAD')
}
