% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_volcano.R
\name{plot_met_volcano}
\alias{plot_met_volcano}
\title{Visualize metastatic propnesity odds ratio in a volcano plot fashion.}
\usage{
plot_met_volcano(x, tumor_type)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}

\item{tumor_type}{The tumour type for which classified data is available. If 'PANCA'
it pools from multiple tumour types into a pan-cancer visualisation.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize metastatic propnesity odds ratio in a volcano plot fashion.
}
\examples{
# First load example classified data
data(MSK_classified)
# Estimate the metastatic propensity associated with mutant TP53, PIK3CA and CDH1 with vs without CNA in BRCA.
for(g in c('TP53', 'PIK3CA', 'CDH1')){MSK_classified = met_propensity(x = MSK_classified, tumor_type = 'BRCA', gene = g)}
# Plot results in a volcano plot
plot_met_volcano(x = MSK_classified, tumor_type = 'BRCA')
}
