% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_k_m.R
\name{plot_posterior_k_m}
\alias{plot_posterior_k_m}
\title{Visualise the posterior distribution on (k,m) configurations.}
\usage{
plot_posterior_k_m(x, k_max = NULL, z_km = NULL)
}
\arguments{
\item{x}{An object of class INCOMMON.}

\item{k_max}{The maximum allowed value of total copyu number.}

\item{z_km}{The list of posterior distributions over (k,m) configurations.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualise the posterior distribution on (k,m) configurations.
}
\examples{
# First load example classified data
data(MSK_genomic_data)
data(MSK_clinical_data)
data(cancer_gene_census)
data(priors_eta)
data(priors_k_m)
x = init(genomic_data = MSK_genomic_data, clinical_data = MSK_clinical_data,  gene_roles = cancer_gene_census)
x = subset_sample(x, "P-0028912")
x = classify(x = x, num_cores = 1, iter_warmup = 10, iter_sampling = 20, num_chains = 1)
# Plot classification results for a specific sample
plot_posterior_k_m(x = x, sample = "P-0028912", purity_error = 0.05)
}
