% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_analysis.R
\name{plot_survival_analysis}
\alias{plot_survival_analysis}
\title{Generate a multipanel figure with Kaplan-Meier survival curves
and forest plot of the Multivariate Cox regression, based on INCOMMON classes.}
\usage{
plot_survival_analysis(
  x,
  tumor_type,
  gene,
  cox_covariates = c("age", "sex", "tmb")
)
}
\arguments{
\item{x}{A list of objects of class \code{'INCOMMON'} containing the classification results for
multiple samples, as produced by using function \code{classify}.}
}
\value{
An object or a list of class \code{'ggplot2'} showing Kaplan-Meier curves and
Cox regression forest plot.
}
\description{
Generate a multipanel figure with Kaplan-Meier survival curves
and forest plot of the Multivariate Cox regression, based on INCOMMON classes.
}
