% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tapacloth.R
\name{plot_tapacloth}
\alias{plot_tapacloth}
\title{Plot results of classification and statistical test.}
\usage{
plot_tapacloth(fit, target_gene)
}
\arguments{
\item{fit}{A fit object as output from \code{analyse_sample()} function.}

\item{target_gene}{The target gene for which classification of mutation calls is wanted.}
}
\value{
A figure assembly including (top) the VAF spectrum calls with expected clonal
peak, purity and VAF of mutations affecting the target gene are highlighted; and (bottom)
a plot showing details of the used statistical test.
}
\description{
Plot results of classification and statistical test.
}
