% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test.R
\name{plot_test}
\alias{plot_test}
\title{Visualize classification results for individual observations.}
\usage{
plot_test(x, model, assembly = F)
}
\arguments{
\item{x}{An object of class \code{'TAPACLOTH'} containing the classification results, as
produced by function \code{run_classifier}.}

\item{model}{Model used for the classification task.}
}
\value{
An object of class \code{'ggplot2'}.
}
\description{
Visualize classification results for individual observations.
}
\examples{
x = init(mutations = example_data$data,
         sample = example_data$sample,
         purity = example_data$purity)
model = "beta-binomial"
x = run_classifier(
    x = x, 
    model = model, 
    rho = 0.01,
    karyotypes = c("1:0","1:1","2:0","2:1","2:2")
    )
plot_test(x = x,id = x$classifier[[model]]data$id[1],model = model)
}
