% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tropism.R
\name{plot_tropism}
\alias{plot_tropism}
\title{Visualize metastatic propnesity odds ratio in a volcano plot fashion.}
\usage{
plot_tropism(x, tumor_type)
}
\arguments{
\item{x}{An object of class \code{'INCOMMON'} containing the classification results, as
produced by function \code{classify}.}

\item{tumor_type}{The tumour type for which classified data is available.}
}
\value{
An object or a list of objects of class \code{'ggplot2'}.
}
\description{
Visualize metastatic propnesity odds ratio in a volcano plot fashion.
}
\examples{
# First load example classified data
data(MSK_PAAD_output)
# Estimate the metastatic tropism associated with mutant TP53, KRAS, CDKN2A, high vs balanced mutant dosage.
for(g in c('TP53', 'KRAS', 'CDKN2A')){for(m in c('Liver', 'Lymph', 'Bone', 'CNS/Brain')){MSK_PAAD_output = met_tropism(x = MSK_PAAD_output, gene = g, tumor_type = 'PAAD', metastatic_site = m)}}
# Plot results in a volcano plot
plot_tropism(x = MSK_PAAD_output, tumor_type = 'PAAD')
}
