% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior}
\alias{posterior}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
posterior(x, id)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{id}{An identifier for the mutation as created by function \code{idify}.}
}
\value{
A table showing posterior distribution and entropy.
}
\description{
Get the model posterior distribution of a mutation.
}
\examples{
# First load example classified data
data(MSK_classified)
# Get posterior distribution of a specific mutation using sample:chr:from:to:ref:alt as id
posterior(MSK_classified, id = 'P-0028912-T01-IM6:chr17:7577121:7577121:G:A:133:837')
}
