% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior}
\alias{posterior}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
posterior(x, id)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}
}
\value{
A table showing posterior distribution and entropy.
}
\description{
Get the model posterior distribution of a mutation.
}
