% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.INCOMMON.R
\name{print.INCOMMON}
\alias{print.INCOMMON}
\title{Print for class \code{'INCOMMON'}.}
\usage{
\method{print}{INCOMMON}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'INCOMMON'}.
}
\examples{
# Example input data from the MSK-MetTropism cohort, released with the package
data(MSK_genomic_data)
print(MSK_genomic_data)
data(MSK_clinical_data)
print(MSK_clinical_data)
# Initialize the INCOMMON object (note the outputs to screen)
x = init(genomic_data = MSK_genomic_data, clinical_data = MSK_clinical_data)
print(x)
}
