% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.INCOMMON.R
\name{print.INCOMMON}
\alias{print.INCOMMON}
\title{Print for class \code{'INCOMMON'}.}
\usage{
\method{print}{INCOMMON}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'INCOMMON'}.
}
