% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.INCOMMON.R
\name{print.INCOMMON}
\alias{print.INCOMMON}
\title{Print for class \code{'INCOMMON'}.}
\usage{
\method{print}{INCOMMON}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'INCOMMON'}.
}
\examples{
x = init(mutations = example_data$data,
         sample = example_data$sample,
         purity = example_data$purity,
         tumor_type = example_data$tumor_type)
print(x)
x = classify(
    x = x, 
    priors = NULL,
    entropy_cutoff = 0.2,
    rho = 0.01,
    karyotypes = c("1:0","1:1","2:0","2:1","2:2")
    )
print(x)
}
