% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{priors}
\alias{priors}
\title{Getter for class \code{'INCOMMON'}.}
\usage{
priors(x)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}
}
\value{
A dplyr::tibble containing prior distributions.
}
\description{
Get model priors used for classification.
}
\examples{
# First load example classified data
data(MSK_classified)
# Get priors used in classification
priors(MSK_classified)
}
