% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/utils.R
\docType{data}
\name{priors_k_m}
\alias{priors_k_m}
\title{Priors from PCAWG}
\format{
A data frame with 4466 rows and 4 columns:
\describe{
\item{gene}{Name of the gene (Hugo Symbol).}
\item{tumor_type}{Tumor type.}
\item{label}{INCOMMON class (\if{html}{\out{<total number of copies>}}N (Mutated: \if{html}{\out{<mutation multiplicity>}}N)).}
\item{p}{Gene and tumor type specific prior probability.}
}
}
\source{
Validated copy number calls from PCAWG: https://doi.org/10.5281/zenodo.6410935
}
\usage{
data(priors_k_m)

priors_k_m(x)
}
\arguments{
\item{x}{An obj of class \code{'INCOMMON'}.}
}
\value{
A dplyr::tibble containing prior distributions.
}
\description{
Prior distribution of copy number and mutation multiplicity from PCAWG.

Get model priors used for classification.
}
\examples{
data(priors_k_m)
priors_k_m
# First load example classified data
data(MSK_PAAD_output)
# Get priors used in classification
priors(MSK_PAAD_output)
}
\keyword{datasets}
