% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{run_classifier}
\alias{run_classifier}
\title{Classify mutations using a (Beta-)Binomial model-based test.}
\usage{
run_classifier(x, alpha_level = 0.01, model = "Binomial", rho = NA)
}
\arguments{
\item{x}{An object of class \code{'TAPACLOTH'} generated with function \code{init}.}

\item{alpha_level}{The significance level to be used in hypothesis testing.}

\item{model}{Model used for the classification task, either "Binomial" (no over-dispersion)
or "Beta-Binomial" (over-dispersion included), that will be used as the expected
distribution for the number of reads with variant at fixed coverage and purity.}

\item{rho}{If "Beta-Binomial" model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the test.}
}
\value{
An object of class \code{TAPACLOTH} containing the input plus
the classification data and paramters.
}
\description{
Classify mutations using a (Beta-)Binomial model-based test.
}
\examples{
x = init(mutations = example_data$data,
         sample = example_data$sample,
         purity = example_data$purity)
x = run_classifier(
    x, 
    alpha_level = 1e-3, 
   model = "Binomial")
print(x)
}
