% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sample.R, R/run_classifier.R
\name{run_classifier}
\alias{run_classifier}
\title{Classify mutations. The classification task can be run either per sample,
using either a Binomial or Beta-Binomial model-based statistical test, or
per gene, in which case the classification is based on a terzile test across
all samples.}
\usage{
run_classifier(data, alpha_level = 0.01, model = "Binomial", rho = NA)

run_classifier(data, alpha_level = 0.01, model = "Binomial", rho = NA)
}
\arguments{
\item{data}{A tibble containing mutations with sample name (sample), gene name
(gene), number of reads with variant (nv), coverage (dp), variant allele
frequency (VAF), and sample purity (purity) as columns.}

\item{alpha_level}{The significance level to be used in hypothesis testing.}

\item{model}{If set to "Binomial" or "Beta-Binomial": classification is run per
sample, using a Binomial (no over-dispersion) or Beta-Binomial (over-dispersion
included) as expected distribution for the number of reads with variant at
fixed coverage and purity. If set to "terzile": classification is run per gene,
and is based on a terzile test.}

\item{rho}{If Beta-Binomial model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the test.}

\item{sample_name}{The name of the sample to be analysed.}
}
\value{
A list containing the input table reduced to the selected sample, with
the additional 'class' column accounting for the classification, the chosen
model, over-dispersion parameter (NA if the model is not Beta-Binomial) and
significance level used in the test.

A list containing the input table reduced to the selected sample, with
the additional 'class' column accounting for the classification, the chosen
model, over-dispersion parameter (NA if the model is not Beta-Binomial) and
significance level used in the test.
}
\description{
Classify mutations. The classification task can be run either per sample,
using either a Binomial or Beta-Binomial model-based statistical test, or
per gene, in which case the classification is based on a terzile test across
all samples.

Classify mutations. The classification task can be run either per sample,
using either a Binomial or Beta-Binomial model-based statistical test, or
per gene, in which case the classification is based on a terzile test across
all samples.
}
