% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classifier.R
\name{run_classifier}
\alias{run_classifier}
\title{Classify mutations per sample using a (Beta-)Binomial model-based test, or
per gene using a 3-quantile test.}
\usage{
run_classifier(x, alpha_level = 0.01, model = "Binomial", rho = NA)
}
\arguments{
\item{alpha_level}{The significance level to be used in hypothesis testing.}

\item{model}{If set to "Binomial" or "Beta-Binomial": classification is run per
sample, using a Binomial (no over-dispersion) or Beta-Binomial (over-dispersion
included) as expected distribution for the number of reads with variant at
fixed coverage and purity. If set to "terzile": classification is run per gene,
and is based on a terzile test.}

\item{rho}{If Beta-Binomial model is selected, this parameter tunes the over-dispersion
of the expected distribution used for the test.}

\item{data}{A tibble containing mutations with sample name (sample), gene name
(gene), number of reads with variant (nv), coverage (dp), variant allele
frequency (VAF), and sample purity (purity) as columns.}
}
\value{
An object of class \code{TAPACLOTH} that represents the classified input data.
}
\description{
Classify mutations per sample using a (Beta-)Binomial model-based test, or
per gene using a 3-quantile test.
}
